+++ ovn_k8s/watcher/watcher.py
@@ -2,13 +2,10 @@ import logging
 import subprocess
 import sys
 
-import eventlet
-eventlet.monkey_patch()
 from oslo_config import cfg
 from oslo_log import log
 import requests
 
-LOGFILE = 'k8s_ovn_watcher.log'
 LOG = log.getLogger(__name__)
 
 K8S_POD_NAME = 'k8s_pod_name'
@@ -181,32 +178,3 @@ def k8s_ns_watcher():
 
 def k8s_nw_policy_watcher():
     LOG.info("Monitoring Kubernetes Network Policies")
-
-
-def _init_conf():
-    # Register options
-    watcher_opts = [
-        cfg.StrOpt('k8s_api_server_host', default='192.168.0.54'),
-        cfg.IntOpt('k8s_api_server_port', default='8080')]
-    cfg.CONF.register_opts(watcher_opts)
-    cfg.CONF(args=sys.argv[1:], project='k8s-ovn-watcher')
-
-
-def main():
-    log.register_options(cfg.CONF)
-    _init_conf()
-    cfg.CONF.set_override('log_file', LOGFILE)
-    cfg.CONF.set_override('debug', True)
-    log.setup(cfg.CONF, 'k8s_ovn_watcher')
-    LOG.info("Kubernetes-OVN watcher process started")
-    cfg.CONF.log_opt_values(LOG, logging.DEBUG)
-    pool = eventlet.greenpool.GreenPool()
-    pool.spawn(ovn_watcher)
-    pool.spawn(k8s_ns_watcher)
-    pool.spawn(k8s_nw_policy_watcher)
-    pool.waitall()
-    LOG.info("Kubernetes-OVN watcher terminated")
-
-
-if __name__ == '__main__':
-    main()
