+++ requirements.txt
@@ -1,0 +1,3 @@
+oslo.config
+oslo.log
+netaddr>=0.7.18
