+++ setup.cfg
@@ -1,0 +1,23 @@
+[metadata]
+name = ovn-kubernetes
+summary = OVN integration for Kubernetes
+description-file = README.md
+author = OVN
+author-email = tbd@idk.org
+classifier =
+    Intended Audience :: Information Technology
+    Intended Audience :: System Administrators
+    License :: OSI Approved :: Apache Software License
+    Operating System :: POSIX :: Linux
+    Programming Language :: Python
+    Programming Language :: Python :: 2
+    Programming Language :: Python :: 2.7
+
+[files]
+packages = ovn_k8s
+scripts = 
+    bin/ovn_cni.py
+
+[entry_points]
+console_scripts =
+    ovn_k8s_watcher = ovn_k8s.cmd.watcher:main
